Split ElaboratorListener implementation into corresponding header and implementaiton files. Also, moved class specific implementation in clone_tree.cpp into its correspondnig ElaboratorListener.cpp file.
