+++ CMakeLists.txt
@@ -145,6 +145,7 @@ add_custom_target(GenerateCode DEPENDS ${model-GENERATED_SRC})
 
 if (WITH_PYTHON_GENERATOR)
   set(uhdm-GENERATED_SRC
+      ${GENDIR}/src/ElaboratorListener.cpp
       ${GENDIR}/src/ExprEval.cpp
       ${GENDIR}/src/Serializer.cpp
       ${GENDIR}/src/Serializer_restore.cpp
@@ -164,6 +165,7 @@ if (WITH_PYTHON_GENERATOR)
   endforeach()
 else()
   set(uhdm-GENERATED_SRC
+      ${GENDIR}/src/ElaboratorListener.cpp
       ${GENDIR}/src/ExprEval.cpp
       ${GENDIR}/src/Serializer.cpp
       ${GENDIR}/src/Serializer_restore.cpp
