+++ scripts/ElaboratorListener_cpp.py
@@ -180,12 +180,12 @@ def generate(models):
     module_listeners = _generate_module_listeners(models)
     class_listeners = _generate_class_listeners(models)
 
-    with open(config.get_template_filepath('ElaboratorListener.h'), 'r+t') as strm:
+    with open(config.get_template_filepath('ElaboratorListener.cpp'), 'r+t') as strm:
         file_content = strm.read()
 
     file_content = file_content.replace('<MODULE_ELABORATOR_LISTENER>', (' ' * 10) + ('\n' + (' ' * 10)).join(module_listeners))
     file_content = file_content.replace('<CLASS_ELABORATOR_LISTENER>', (' ' * 4) + ('\n' + (' ' * 4)).join(class_listeners))
-    file_utils.set_content_if_changed(config.get_output_header_filepath('ElaboratorListener.h'), file_content)
+    file_utils.set_content_if_changed(config.get_output_source_filepath('ElaboratorListener.cpp'), file_content)
 
     return True
