+++ templates/clone_tree.cpp
@@ -36,142 +36,6 @@ BaseClass* clone_tree(const BaseClass* root, Serializer& s,
   return root ? root->DeepClone(&s, elaborator, nullptr) : nullptr;
 }
 
-// Hardcoded implementations
-
-any* ElaboratorListener::bindNet(const std::string& name) {
-  for (InstStack::reverse_iterator i = instStack_.rbegin();
-       i != instStack_.rend(); ++i) {
-    ComponentMap& netMap = std::get<0>((*i).second);
-    ComponentMap::iterator netItr = netMap.find(name);
-    if (netItr != netMap.end()) {
-      return (any*)(*netItr).second;
-    }
-  }
-  return nullptr;
-}
-
-// Bind to a net or parameter in the current instance
-any* ElaboratorListener::bindAny(const std::string& name) {
-  for (InstStack::reverse_iterator i = instStack_.rbegin();
-       i != instStack_.rend(); ++i) {
-    ComponentMap& netMap = std::get<0>((*i).second);
-    ComponentMap::iterator netItr = netMap.find(name);
-    if (netItr != netMap.end()) {
-      return (any*)(*netItr).second;
-    }
-
-    ComponentMap& paramMap = std::get<1>((*i).second);
-    ComponentMap::iterator paramItr = paramMap.find(name);
-    if (paramItr != paramMap.end()) {
-      return (any*)(*paramItr).second;
-    }
-  }
-  return nullptr;
-}
-
-// Bind to a param in the current instance
-any* ElaboratorListener::bindParam(const std::string& name) {
-  for (InstStack::reverse_iterator i = instStack_.rbegin();
-       i != instStack_.rend(); ++i) {
-    ComponentMap& paramMap = std::get<1>((*i).second);
-    ComponentMap::iterator paramItr = paramMap.find(name);
-    if (paramItr != paramMap.end()) {
-      return (any*)(*paramItr).second;
-    }
-  }
-  return nullptr;
-}
-
-// Bind to a function or task in the current scope
-any* ElaboratorListener::bindTaskFunc(const std::string& name,
-                                      const class_var* prefix) {
-  for (InstStack::reverse_iterator i = instStack_.rbegin();
-       i != instStack_.rend(); ++i) {
-    ComponentMap& funcMap = std::get<2>((*i).second);
-    ComponentMap::iterator funcItr = funcMap.find(name);
-    if (funcItr != funcMap.end()) {
-      return (any*)(*funcItr).second;
-    }
-  }
-  if (prefix) {
-    const typespec* tps = prefix->Typespec();
-    if (tps && tps->UhdmType() == uhdmclass_typespec) {
-      const class_defn* def = ((class_typespec*)tps)->Class_defn();
-      while (def) {
-        if (def->Task_funcs()) {
-          for (task_func* tf : *def->Task_funcs()) {
-            if (tf->VpiName() == name) return tf;
-          }
-        }
-        const UHDM::extends* ext = def->Extends();
-        if (ext) {
-          const class_typespec* tps = ext->Class_typespec();
-          def = tps->Class_defn();
-        } else {
-          break;
-        }
-      }
-    }
-  }
-  return nullptr;
-}
-
-bool ElaboratorListener::isFunctionCall(const std::string& name,
-                                        const expr* prefix) {
-  if (instStack_.size()) {
-    for (InstStack::reverse_iterator i = instStack_.rbegin();
-         i != instStack_.rend(); ++i) {
-      ComponentMap& funcMap = std::get<2>((*i).second);
-      ComponentMap::iterator funcItr = funcMap.find(name);
-      if (funcItr != funcMap.end()) {
-        return ((*funcItr).second->UhdmType() == uhdmfunction);
-      }
-    }
-  }
-  if (prefix) {
-    const ref_obj* ref = any_cast<const ref_obj*>(prefix);
-    const class_var* vprefix = nullptr;
-    if (ref) vprefix = any_cast<const class_var*>(ref->Actual_group());
-    any* func = bindTaskFunc(name, vprefix);
-    if (func) {
-      if (func->UhdmType() == uhdmfunction) {
-        return true;
-      } else {
-        return false;
-      }
-    }
-  }
-  return true;
-}
-
-bool ElaboratorListener::isTaskCall(const std::string& name,
-                                    const expr* prefix) {
-  if (instStack_.size()) {
-    for (InstStack::reverse_iterator i = instStack_.rbegin();
-         i != instStack_.rend(); ++i) {
-      ComponentMap& funcMap = std::get<2>((*i).second);
-      ComponentMap::iterator funcItr = funcMap.find(name);
-      if (funcItr != funcMap.end()) {
-        return ((*funcItr).second->UhdmType() == uhdmtask);
-      }
-    }
-  }
-  if (prefix) {
-    const ref_obj* ref = any_cast<const ref_obj*>(prefix);
-    const class_var* vprefix = nullptr;
-    if (ref) vprefix = any_cast<const class_var*>(ref->Actual_group());
-    any* func = bindTaskFunc(name, vprefix);
-    if (func) {
-      if (func->UhdmType() == uhdmtask) {
-        return true;
-      } else {
-        return false;
-      }
-    }
-  }
-  return true;
-}
-
 tf_call* sys_func_call::DeepClone(Serializer* serializer,
                                   ElaboratorListener* elaborator,
                                   BaseClass* parent) const {
@@ -852,188 +716,6 @@ task* task::DeepClone(Serializer* serializer, ElaboratorListener* elaborator,
   return clone;
 }
 
-static void propagateParamAssign(param_assign* pass, const any* target) {
-  UHDM_OBJECT_TYPE targetType = target->UhdmType();
-  Serializer& s = *pass->GetSerializer();
-  switch (targetType) {
-    case uhdmclass_defn: {
-      class_defn* defn = (class_defn*)target;
-      const any* lhs = pass->Lhs();
-      const std::string& name = lhs->VpiName();
-      VectorOfany* params = defn->Parameters();
-      if (params) {
-        for (any* param : *params) {
-          if (param->VpiName() == name) {
-            VectorOfparam_assign* passigns = defn->Param_assigns();
-            if (passigns == nullptr) {
-              defn->Param_assigns(s.MakeParam_assignVec());
-              passigns = defn->Param_assigns();
-            }
-            param_assign* pa = s.MakeParam_assign();
-            pa->Lhs(param);
-            pa->Rhs((any*)pass->Rhs());
-            passigns->push_back(pa);
-          }
-        }
-      }
-      const UHDM::extends* extends = defn->Extends();
-      if (extends) {
-        propagateParamAssign(pass, extends->Class_typespec());
-      }
-      const auto vars = defn->Variables();
-      if (vars) {
-        for (auto var : *vars) {
-          propagateParamAssign(pass, var);
-        }
-      }
-      break;
-    }
-    case uhdmclass_var: {
-      class_var* var = (class_var*)target;
-      propagateParamAssign(pass, var->Typespec());
-      break;
-    }
-    case uhdmclass_typespec: {
-      class_typespec* defn = (class_typespec*)target;
-      const any* lhs = pass->Lhs();
-      const std::string& name = lhs->VpiName();
-      VectorOfany* params = defn->Parameters();
-      if (params) {
-        for (any* param : *params) {
-          if (param->VpiName() == name) {
-            VectorOfparam_assign* passigns = defn->Param_assigns();
-            if (passigns == nullptr) {
-              defn->Param_assigns(s.MakeParam_assignVec());
-              passigns = defn->Param_assigns();
-            }
-            param_assign* pa = s.MakeParam_assign();
-            pa->Lhs(param);
-            pa->Rhs((any*)pass->Rhs());
-            passigns->push_back(pa);
-          }
-        }
-      }
-      const class_defn* def = defn->Class_defn();
-      if (def) {
-        propagateParamAssign(pass, (class_defn*)def);
-      }
-      break;
-    }
-    default:
-      break;
-  }
-}
-
-void ElaboratorListener::enterVariables(const variables* object,
-                                        const BaseClass* parent,
-                                        vpiHandle handle,
-                                        vpiHandle parentHandle) {
-  Serializer* s = ((variables*)object)->GetSerializer();
-  if (object->UhdmType() == uhdmclass_var) {
-    if (!inHierarchy_)
-      return;  // Only do class var propagation while in elaboration
-    const class_var* cv = (class_var*)object;
-    class_var* const rw_cv = (class_var*)cv;
-    class_typespec* ctps = (class_typespec*)cv->Typespec();
-    if (ctps) {
-      ctps = ctps->DeepClone(s, this, rw_cv);
-      rw_cv->Typespec(ctps);
-      VectorOfparam_assign* params = ctps->Param_assigns();
-      if (params) {
-        for (param_assign* pass : *params) {
-          propagateParamAssign(pass, ctps->Class_defn());
-        }
-      }
-    }
-  }
-}
-
-void ElaboratorListener::leaveVariables(const variables* object,
-                                        const BaseClass* parent,
-                                        vpiHandle handle,
-                                        vpiHandle parentHandle) {}
-
-void ElaboratorListener::enterTask_func(const task_func* object,
-                                        const BaseClass* parent,
-                                        vpiHandle handle,
-                                        vpiHandle parentHandle) {
-  // Collect instance elaborated nets
-  ComponentMap varMap;
-  if (object->Variables()) {
-    for (variables* var : *object->Variables()) {
-      varMap.insert(std::make_pair(var->VpiName(), var));
-    }
-  }
-  if (object->Io_decls()) {
-    for (io_decl* decl : *object->Io_decls()) {
-      varMap.insert(std::make_pair(decl->VpiName(), decl));
-    }
-  }
-  varMap.insert(std::make_pair(object->VpiName(), object->Return()));
-
-  ComponentMap paramMap;
-
-  ComponentMap funcMap;
-
-  instStack_.push_back(
-      std::make_pair(object, std::make_tuple(varMap, paramMap, funcMap)));
-}
-
-void ElaboratorListener::leaveTask_func(const task_func* object,
-                                        const BaseClass* parent,
-                                        vpiHandle handle,
-                                        vpiHandle parentHandle) {
-  instStack_.pop_back();
-}
-
-void ElaboratorListener::enterGen_scope(const gen_scope* object,
-                                        const BaseClass* parent,
-                                        vpiHandle handle,
-                                        vpiHandle parentHandle) {
-  // Collect instance elaborated nets
-
-  ComponentMap netMap;
-  if (object->Nets()) {
-    for (net* net : *object->Nets()) {
-      netMap.insert(std::make_pair(net->VpiName(), net));
-    }
-  }
-  if (object->Array_nets()) {
-    for (array_net* net : *object->Array_nets()) {
-      netMap.insert(std::make_pair(net->VpiName(), net));
-    }
-  }
-  // Collect instance parameters, defparams
-  ComponentMap paramMap;
-  if (object->Parameters()) {
-    for (any* param : *object->Parameters()) {
-      paramMap.insert(std::make_pair(param->VpiName(), param));
-    }
-  }
-  if (object->Def_params()) {
-    for (def_param* param : *object->Def_params()) {
-      paramMap.insert(std::make_pair(param->VpiName(), param));
-    }
-  }
-  if (object->Variables()) {
-    for (variables* var : *object->Variables()) {
-      paramMap.insert(std::make_pair(var->VpiName(), var));
-    }
-  }
-
-  ComponentMap funcMap;
-
-  instStack_.push_back(
-      std::make_pair(object, std::make_tuple(netMap, paramMap, funcMap)));
-}
-
-void ElaboratorListener::leaveGen_scope(const gen_scope* object,
-                                        const BaseClass* parent,
-                                        vpiHandle handle,
-                                        vpiHandle parentHandle) {
-  instStack_.pop_back();
-}
-
 cont_assign* cont_assign::DeepClone(Serializer* serializer,
                                     ElaboratorListener* elaborator,
                                     BaseClass* parent) const {
@@ -1171,13 +853,6 @@ hier_path* hier_path::DeepClone(Serializer* serializer,
   }
 
   return clone;
-}
-
-void ElaboratorListener::leaveRef_obj(const ref_obj* object,
-                                      const BaseClass* parent, vpiHandle handle,
-                                      vpiHandle parentHandle) {
-  if (!object->Actual_group())
-    ((ref_obj*)object)->Actual_group(bindAny(object->VpiName()));
 }
 
 // Auto generated implementations
